Fix invoice generation for newly-started SIMs.
Their monthly cost item lines would have the wrong dates for the start and end dates of the monthly cost (i.e. 29-01-2013 to 19-02-2013 instead of 01-02-2013 to 28-02-2013). This was caused by the contractStartDate being in a local time-zone instead of in UTC. Also, now we force the contract to start on the first day of the next month, just like SpeakUp does.