Include queued item lines on an invoice.
If an invoice had a total price of 0, it might still have item lines that were fully matched by "negative" queued item lines. So check the number of item lines instead of the total price. Return 204 No Content if there were no item lines, but happily store and return an invoice with a total of 0.